+++ packages/orm/angel_orm_postgres/example/main.dart
@@ -1,12 +1,20 @@
 import 'dart:io';
 import 'package:angel3_orm_postgres/angel3_orm_postgres.dart';
-import 'package:postgres/postgres.dart';
+import 'package:postgres_pool/postgres_pool.dart';
 
 void main() async {
-  var executor = PostgreSqlExecutorPool(Platform.numberOfProcessors, () {
-    return PostgreSQLConnection('localhost', 5432, 'orm_test',
-        username: 'test', password: 'test123');
-  });
+  var executor = PostgreSqlPoolExecutor(PgPool(
+    PgEndpoint(
+      host: 'localhost',
+      port: 5432,
+      database: 'orm_test',
+      username: Platform.environment['POSTGRES_USERNAME'] ?? 'test',
+      password: Platform.environment['POSTGRES_PASSWORD'] ?? 'test123',
+    ),
+    settings: PgPoolSettings()
+      ..maxConnectionAge = Duration(hours: 1)
+      ..concurrency = 5,
+  ));
 
   var rows = await executor.query('users', 'SELECT * FROM users', {});
   print(rows);