+++ packages/orm/angel_orm_test/lib/src/many_to_many_test.dart
@@ -21,7 +21,9 @@ void manyToManyTests(FutureOr<QueryExecutor> Function() createExecutor,
     //var rows = await executor.query(null, query, {});
     var rows = await executor.query('', query, {});
     print('\n${rows.length} row(s):');
-    rows.forEach((r) => print('  * $r'));
+    for (var r in rows) {
+      print('  * $r');
+    }
     print('==================================================\n\n');
   }