+++ packages/orm/angel_orm_test/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 3.0.3
 
-* Added performance test cases
+* Added `performance_test` test cases
+* Update `edge_case_test` test case to support connection pooling
 
 ## 3.0.2